requirements.txt
fyers-apiv3
pandas
pandas_ta
python-dotenv
pyTelegramBotAPI
